+++ src/powerglove_vision/__init__.py
@@ -11,4 +11,4 @@
 
 """PowerGlove Vision camera controller."""
 
-__version__ = "0.2.0.dev2"
+__version__ = "0.2.0.dev3"